+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Run Configurations:
 
 1.  Fixed scrolling to an item in the results pane when the detected filetype
     is plain text.
+1.  Fixed a null pointer exception when processing MarkLogic syntax error
+    responses.
 
 ## 2020